Fixed gcc compilation warnings
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@3852 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ include/gpac/internal/mpd.h
@@ -217,9 +217,9 @@ typedef struct
 	GF_MPD_COMMON_ATTRIBUTES_ELEMENTS
 
 	/*if not set in MPD, an ID is dynamically generated (u32 cast of the structure)*/
-	u32 id;
+	u64 id;
 	/*if not set in MPD, an ID is dynamically generated (u32 cast of the structure)*/
-	u32 group;
+	u64 group;
 
 	char *lang;
 	char *content_type;
